# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PAOSITRA - API de gestion",
    "description": "API versionnée des lots Trésorerie et Gestion des opérations. Les choix techniques documentés ne constituent pas des exigences contractuelles du DAO.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/treasury/attachments/objects/{objectType}/{objectId}": {
      "get": {
        "tags": [
          "Tresorerie"
        ],
        "summary": "Liste des pieces jointes d'un objet metier",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pieces jointes actives"
          }
        }
      },
      "post": {
        "tags": [
          "Tresorerie"
        ],
        "summary": "Depot securise d'une piece jointe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Piece enregistree"
          }
        }
      }
    },
    "/api/v1/treasury/attachments/file/{id}": {
      "get": {
        "tags": [
          "Tresorerie"
        ],
        "summary": "Telechargement controle d'une piece jointe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Contenu du fichier"
          }
        }
      }
    },
    "/api/v1/treasury/attachments/file/{id}/archive": {
      "post": {
        "tags": [
          "Tresorerie"
        ],
        "summary": "Archivage logique d'une piece jointe",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Piece archivee"
          }
        }
      }
    },
...
  }
}